+++ src/components/ui/card/index.jsx
@@ -9,6 +9,7 @@ import styled from 'styled-components';
 import Divider from '@mui/material/Divider';
 import DownloadIcon from '@mui/icons-material/Download';
 import './index.css';
+import { useConfig } from '../../../context/configContext';
 
 const HorizontalLayout = styled.div`
   width: 100%;
@@ -90,10 +91,11 @@ export const VisualizationItemCard = forwardRef(
       onSelectVizLayer,
       onHoverOnVizLayer,
       hoveredVizItemId,
-      rasterApiUrl,
     },
     ref
   ) => {
+    const { config } = useConfig();
+    const rasterApiUrl = config.rasterApiUrl;
     const vizItemSourceId = vizItem?.id;
     const orbit = vizItem?.plumeProperties?.orbit;
     const imageUrl = `${rasterApiUrl}/collections/emit-ch4plume-v1/items/${vizItemSourceId}/preview.png?bidx=1&assets=ch4-plume-emissions&rescale=1%2C1500&resampling=bilinear&colormap_name=plasma`;
