+++ src/pages/dashboard/helper/PlumeLayer.jsx
@@ -29,7 +29,6 @@ function Plumes({
   vizItems,
   highlightedLayer,
   onHoverOverLayer,
-  config,
 }) {
   const { map } = useMapbox();
 
@@ -128,7 +127,6 @@ function Plumes({
       highlightedLayer={highlightedLayer}
       onClickedOnLayer={handleClickedOnLayer}
       handleRemoveLayer={handleRemoveLayer}
-      config={config}
     />
   );
 }
