+++ src/pages/dashboard/index.jsx
@@ -16,7 +16,7 @@ import {
   CoverageLayers,
   MapViewPortComponent,
   RASTER_ZOOM_LEVEL,
-  ToggleSwitch
+  ToggleSwitch,
 } from '../../components';
 
 import styled from 'styled-components';
@@ -25,6 +25,7 @@ import './index.css';
 import { filterByDateRange, getPopupContent } from './helper';
 import Plumes from './helper/PlumeLayer';
 import moment from 'moment';
+import { useConfig } from '../../context/configContext';
 
 const TITLE = 'EMIT Methane Plume Viewer';
 const DESCRIPTION =
@@ -69,7 +70,6 @@ export function Dashboard({
   setZoomLevel,
   collectionId,
   loadingData,
-  config,
 }) {
   // states for data
   const [vizItems, setVizItems] = useState([]); // store all available visualization items
@@ -85,6 +85,8 @@ export function Dashboard({
 
   // states for components/controls
   const [openDrawer, setOpenDrawer] = useState(false);
+
+  const { config } = useConfig();
 
   //colormap states
   const [VMAX, setVMAX] = useState(100);
@@ -205,7 +207,7 @@ export function Dashboard({
   return (
     <div className='fullSize'>
       <div id='dashboard-map-container'>
-        <MainMap config={config}>
+        <MainMap>
           <Paper className='title-container'>
             <Title title={TITLE} description={DESCRIPTION} />
             <div className='title-content'>
@@ -275,7 +277,6 @@ export function Dashboard({
             assets={assets}
             onHoverOverLayer={handleHoveredVizLayer}
             highlightedLayer={hoveredVizLayerId}
-            config={config}
           />
           )
         </MainMap>
@@ -288,7 +289,6 @@ export function Dashboard({
           collectionId={collectionId}
           onSelectVizLayer={handleClickedVizLayer}
           onHoverOnVizLayer={handleHoveredVizLayer}
-          config={config}
         />
         )
       </div>
