+++ src/pages/emitInterface/types.ts
@@ -17,23 +17,5 @@ export interface EmitInterfaceConfig {
 }
 
 export interface EmitInterfaceProps {
-  /**
-   * The STAC collection ID to fetch data from
-   */
-  collectionId?: string;
-
-  /**
-   * Initial zoom location [longitude, latitude]
-   */
-  zoomLocation?: [number, number];
-
-  /**
-   * Initial zoom level
-   */
-  zoomLevel?: number;
-
-  /**
-   * Configuration object for API endpoints and other settings
-   */
   config?: Partial<EmitInterfaceConfig>;
 }
